feat(package): use explicit fileset for package src
This uses a fileset to restrict the amount of rebuilds that one has
to perform.

For example, if a test is changed, we do not want to rebuild the
`nixos-cli` package, since that's wasteful. Using an explicit opt-in
fileset for src works here because there are only a few files/directories
required to compile the actual package, and they can be specified
manually without much trouble.